# accountingdashboard.client/src/api/models/CreateRuleRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type CreateRuleRequest = {
    client: string;
    program: string;
    depositDestination: string;
};