# accountingdashboard.client/src/api/models/Rule.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type Rule = {
    id?: number;
    client: string;
    program: string;
    depositDestination: string;
    updated?: string;
};