# accountingdashboard.client/src/api/models/UpdateRuleRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type UpdateRuleRequest = {
    client: string;
    program: string;
    depositDestination: string;
};